2010-09-15  Simon Fraser  <simon.fraser@apple.com>
        Reviewed by Sam Weinig.

        <rdar://problem/8214099> Mousewheel scrolling in WebKit2 is wonky

        Use floats so that fractional scrollwheel deltas don't get lost.

        * Shared/mac/WebEventFactory.mm:
        (WebKit::WebEventFactory::createWebWheelEvent):

Canonical link: https://commits.webkit.org/58313@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@67589 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ WebKit2/Shared/mac/WebEventFactory.mm
@@ -987,10 +987,10 @@ WebWheelEvent WebEventFactory::createWebWheelEvent(NSEvent *event, NSView *windo
     int globalPositionY                     = globalPosition.y;
     WebWheelEvent::Granularity granularity  = WebWheelEvent::ScrollByPixelWheelEvent;
 
-    int deltaX = 0;
-    int deltaY = 0;
-    int wheelTicksX = 0;
-    int wheelTicksY = 0;
+    float deltaX = 0;
+    float deltaY = 0;
+    float wheelTicksX = 0;
+    float wheelTicksY = 0;
     if ([event _continuousScroll]) {
         // smooth scroll events
         deltaX = [event deviceDeltaX];
